Переписал классы проверок элементов на странице (#6)
* Added "raw" version of test Administration link Employee

* Edited file LinkEmployee.feature

* Edited scenario

* Check classes rewritten
